project/requirements.txt
torch
torchvision
ultralytics
opencv-python-headless
Pillow
numpy
